fix urls
