+++ components/yur-face-page/CardBlock/CardBlock.tsx
@@ -4,7 +4,7 @@ import Image from "next/image";
 import NextLink from "next/link";
 import { useTranslation } from "react-i18next";
 import { CardBlockProps } from "./CardBlock.t";
-
+import slugify from "slugify";
 import {
   CardBody,
   CradTitle,
@@ -58,7 +58,7 @@ const Card: FC<CardBlockProps> = ({ item }) => {
               </CardServiceText>
             </CardServiceBlock>
             <CardServiceBlock>
-              <CardServiceHeading>{title_2}</CardServiceHeading>
+              <CardServiceHeading>{title_2} </CardServiceHeading>
               <CardServiceText>
               {description_2}
 
@@ -72,11 +72,11 @@ const Card: FC<CardBlockProps> = ({ item }) => {
             </CardServiceBlock>
           </CardService>
           <CardBtns>
-          <Link href={`/mtpl-insurance?id=${id}`} passHref>
+          <Link href={`/insurance/${id}/${slugify(main_title)}`} passHref>
             <CardButton>{t("common:Get_policy")}</CardButton>
           </Link>
             <HorizontalSliderButton>
-              <Link href={`/mtpl-insurance?id=${id}`} passHref>
+              <Link href={`/insurance/${id}/${slugify(main_title)}`} passHref>
                 <a>
                   <span>{t("common:Learn_more")}</span>
                   <ArrowIcon fill="#F0803D" class="arrow-right" />
